removed online indicator in front end
only in the front end, in the database, the accounts still have their active indicator. This is due to the problem that when the user closes their current tab or browser without logging out,  they will still be considered online because it can only be updated to offline when the user logs out. I tried using the socketio or the unload, it works but those are getting triggered even if the user just reloads the page.